Replace old Geo svc with new implementation
