+++ k4ActsTracking/include/k4ActsTracking/IActsGeoSvc.h
@@ -22,6 +22,7 @@
 
 #include <GaudiKernel/IService.h>
 
+#include <cstdint>
 #include <memory>
 #include <unordered_map>
 
@@ -39,13 +40,14 @@ namespace Acts {
 
 class GAUDI_API IActsGeoSvc : virtual public IService {
 public:
-  using VolumeSurfaceMap = std::unordered_map<uint64_t, const Acts::Surface*>;
+  using CellIDSurfaceMap = std::unordered_map<uint64_t, const Acts::Surface*>;
 
 public:
   DeclareInterfaceID(IActsGeoSvc, 1, 0);
 
-  virtual std::shared_ptr<const Acts::TrackingGeometry>      trackingGeometry() const = 0;
-  virtual std::shared_ptr<const Acts::MagneticFieldProvider> magneticField() const    = 0;
+  virtual std::shared_ptr<const Acts::TrackingGeometry>      trackingGeometry() const   = 0;
+  virtual std::shared_ptr<const Acts::MagneticFieldProvider> magneticField() const      = 0;
+  virtual const CellIDSurfaceMap&                            cellIdToSurfaceMap() const = 0;
 
   virtual ~IActsGeoSvc() = default;
 };
